more moving around
